requirements.txt
alice-blue==1.0.1
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.6
beautifulsoup4==4.9.3
boto3==1.17.9
botocore==1.20.9
bs4==0.0.1
bsedata==0.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.6
dateutils==0.6.12
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-heroku==0.3.1
django-redis==4.12.1
django-redis-cache==3.0.0
django-storages==1.11.1
finnhub-python==2.4.0
fyers-api==1.0.9
get-all-tickers==1.7
gunicorn==20.0.4
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
lxml==4.6.2
mccabe==0.6.1
multitasking==0.0.9
nse50bse30==0.0.3
nsepython==0.0.94
nsetools==1.0.11
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.3
pep8==1.7.1
protlib==1.4
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tradingview-ta==3.2.2
urllib3==1.26.3
websocket-client==0.58.0
whitenoise==5.2.0
wrapt==1.12.1
yfinance==0.1.59
